# HolidayDessertStore/bin/Release/net9.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Holiday Dessert Store API",
    "version": "v1"
  },
  "paths": {
    "/api/DessertManagement": {
      "post": {
        "tags": [
          "DessertManagement"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Dessert"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DessertManagement/{id}": {
      "get": {
        "tags": [
          "DessertManagement"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Dessert"
            }
          }
        }
      },
      "put": {
        "tags": [
          "DessertManagement"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Dessert"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "DessertManagement"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "Dessert": {
      "required": [
        "description",
        "isAvailable",
        "name",
        "price",
        "quantity"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "description": {
          "minLength": 1,
          "type": "string"
        },
        "price": {
          "format": "double",
          "maximum": 1000,
          "minimum": 0.01,
          "type": "number"
        },
        "imagePath": {
          "type": "string"
        },
        "quantity": {
          "format": "int32",
          "maximum": 100,
          "minimum": 1,
          "type": "integer"
        },
        "isAvailable": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    }
  }
}